% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_noreps.R
\name{get_stan_summary_noreps}
\alias{get_stan_summary_noreps}
\title{parameters statistical results}
\usage{
get_stan_summary_noreps(nested_stanfit, nested_data)
}
\arguments{
\item{nested_stanfit}{output of rstan::sampling}

\item{nested_data}{input data into rstan::sampling}
}
\value{
A data frame
}
\description{
Get the estimated statistical results of parameters from a stan model
Extracts the names of transcript_IDs from the input data
}
\details{
The input data is the output of the rstan::sampling function
Output a data frame that includes a statistical result of parameters.
}
\keyword{extraction}
\keyword{parameter}
