% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{load_counts}
\alias{load_counts}
\title{Load the count matrix}
\usage{
load_counts(count_file, samplesheet)
}
\arguments{
\item{count_file}{the count matrix file}

\item{samplesheet}{sample description spreadsheet, either a tibble or a file}
}
\description{
Load the count matrix and combine with the samplesheet
}
\details{
In this function, function read_count_files from this package is called.
Output a wider data frame for representing the count value of each transcript
for each fraction.
}
\keyword{counts}
\keyword{each}
\keyword{fraction}
