% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_reps.R
\name{model_fit_reps}
\alias{model_fit_reps}
\title{Model fit}
\usage{
model_fit_reps(nested_data, chains = 4, iter = 1000, ...)
}
\arguments{
\item{nested_data}{output of prepare_data_with_reps}

\item{chains}{A number, defaulting to 4}

\item{iter}{A number, defaulting to 1000}

\item{...}{Other arguments passed to \code{rstan::sampling} (e.g. warmup, seed).}
}
\value{
A dataframe with a column containing the class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Fit the bayesian statistical model to the counts data.
}
\details{
The input data should be from prepare_data_with_reps
}
\seealso{
\code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}, \code{browseVignettes("rstan")}
}
\keyword{fit}
\keyword{model}
