% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_reps.R
\name{prepare_data_with_reps}
\alias{prepare_data_with_reps}
\title{Prepare Data with Replicates}
\usage{
prepare_data_with_reps(count_data, min_counts = 20)
}
\arguments{
\item{count_data}{A data frame containing count data (output of load_counts).}

\item{min_counts}{Minimum count threshold for filtering low counts (default = 20).}
}
\value{
A tibble containing a column of prepared data in the form of a named list:
\itemize{
\item{"NREP"}{Number of replicates.}
\item{"NRNA"}{Number of RNA sequences.}
\item{"tot_obs_counts"}{Total observed counts matrix.}
\item{"sup_obs_counts"}{Supernatant observed counts matrix.}
\item{"pel_obs_counts"}{Pellet observed counts matrix.}
}
}
\description{
This function prepares count data by filtering transcripts with low counts, rounding counts to integers, and
dividing data into nested fractions. It then converts the data to a count matrix of transcript_IDs and Reps.
}
