% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_count_files}
\alias{read_count_files}
\title{Read count files}
\usage{
read_count_files(file)
}
\arguments{
\item{file}{The input count file}
}
\description{
Read a matrix of counts for each sample
}
\details{
This function is to read a count matrix containing a column of transcript_IDs
then a column of counts for every Sample_ID.
}
\keyword{counts}
