% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_samplesheet}
\alias{read_samplesheet}
\title{Read Samplesheet}
\usage{
read_samplesheet(file)
}
\arguments{
\item{file}{The input Samplesheet file name}
}
\value{
A single data frame
}
\description{
Read the Samplesheet file that is provided by users.
}
\details{
This function is to read the samplesheet file that is provided by users.
Spreadsheet has 4 columns with headers: Sample_ID, Condition, Rep, Fraction.
Sample_IDs need to be unique, Fractions should be Total, Supernatant, or Pellet.
}
\keyword{samplesheet}
