% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SedSeqQuant-package.R
\docType{package}
\name{SedSeqQuant-package}
\alias{SedSeqQuant-package}
\alias{SedSeqQuant}
\title{The 'SedSeqQuant' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.22. https://mc-stan.org
}
